load HAM10k nearly done, oversampling for class balance

+++ utils/load_mnist.py
@@ -14,8 +14,8 @@ def prepare_mnist(site):
     
     if site == "test":
         return (
-            x_test[...,np.newaxis].astype(np.float32), 
-            y_test.astype(np.int32),
+            x_test[...,np.newaxis].astype(np.float64), 
+            y_test.astype(np.int64),
         )
         
   
@@ -49,6 +49,6 @@ def prepare_mnist(site):
         
 
     return (
-        x_train[train_idx][...,np.newaxis].astype(np.float32), 
-        y_train[train_idx].astype(np.int32), 
+        x_train[train_idx][...,np.newaxis].astype(np.float64), 
+        y_train[train_idx].astype(np.int64), 
     )
